refactor: split shared test helper types
